Check for existing user before trying to create
Change the logic so that a call is made to check by email
address for an existing Dataverse user and only if the
user is not found then a new authenticated user is provisioned
using a HTTP POST to /api/admin/authenticatedUsers.

Also fix a bug in comparison between the Dataverse authenticated
user ID and the Registry ProvisioningTarget Identifier.